requirements.txt
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.1
pymongo==3.11.3
python-dotenv==0.17.1
regex==2021.3.17
six==1.15.0
tqdm==4.59.0
Werkzeug==1.0.1
